Temporary renamed the main class (easier migration).
